refactor(codex-adapter): carve history.rs into history_convert/history_collab/unified_diff
Todo #247 task 2: same 300-line-ceiling pressure as event_mapper.rs.
Split convert_thread_items into history_convert.rs, emit_collab_agent
into history_collab.rs, and the parse_unified_diff shim into
unified_diff.rs, all verbatim moves with re-exports.
